panfrost: use 64-bits for layout calculations
On modern Mali GPUs, we can have 16 bits for the X and Y sizes, already
overflowing 32-bit barrier even with a single layer of byte-sized
formats.

So let's make sure we have enough bits to avoid overflows here.

Fixes: d5ed77800e0 ("panvk: Fix GetPhysicalDeviceProperties2() to report accurate info")
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/32221>
(cherry picked from commit 00b25ec7693e81c23619369551b0d1fd0cb9c440)

+++ src/panfrost/lib/pan_texture.h
@@ -112,7 +112,7 @@ struct pan_image_layout {
    struct pan_image_slice_layout slices[MAX_MIP_LEVELS];
 
    uint64_t data_size;
-   unsigned array_stride;
+   uint64_t array_stride;
 };
 
 struct pan_image_mem {
